+++ packages/common/src/ee/AiAgent/index.ts
@@ -185,16 +185,19 @@ export type ApiAiAgentThreadResponse = {
     results: AiAgentThread;
 };
 
-export type ApiAiAgentThreadGenerateRequest = {
+export type ApiAiAgentThreadCreateRequest = {
+    prompt?: string;
+};
+
+export type ApiAiAgentThreadCreateResponse = ApiSuccess<AiAgentThreadSummary>;
+
+export type ApiAiAgentThreadMessageCreateRequest = {
     prompt: string;
 };
 
-export type ApiAiAgentThreadGenerateResponse = {
-    status: 'ok';
-    results: {
-        jobId: string;
-    };
-};
+export type ApiAiAgentThreadMessageCreateResponse = ApiSuccess<
+    AiAgentMessageUser<AiAgentUser>
+>;
 
 export type ApiAiAgentStartThreadResponse = {
     status: 'ok';